Refact: changed return from decimal to double

+++ CalculaJurosAPI/Services/ITaxaDeJurosService.cs
@@ -2,6 +2,6 @@
 {
     public interface ITaxaDeJurosService
     {
-        Task<decimal> GetAsync();
+        Task<double> GetAsync();
     }
 }
